drivers: serial: cp210x: backport driver to 5.10.72
Due to iei custom hardware design does not use generic rts direction pin
for RS485 mode, so need adapt an additional gpio as direction pin, but
it's not match geneirc design for new driver in 5.15.71.

cp210x.c in kernel 5.10.72 which is merged from usb-serial-5.1-rc1-2
branch, author is Johan Hovold/

But cp210x.c in kernel 5.15.71, which is merged from
usb-serial-5.12-rc1, Johan Hovold improve much structure for drivers,
but it's not suitable for iei current hardware design, so back port to
usb-serial-5.1-rc1-2 driver, it works after tested.