LF-16555: drm/bridge: imx95-mipi-dsi: Set clkdiv_clk_en bit in MIPI DPHY
For DPHY configurations which use a lane bandwidth < 450Mbps the
clkdiv_clk_en bit should be set as per the DPHY databook. Without this
bit set the MIPI high speed clock from the SoC can be invalid, which
causes the video data to be corrupted.

For a video mode such as 720x480@60 and a 4 lane DPHY configuration the
required per lane bandwidth is 162Mbps. When clkdiv_clk_en is cleared
there is no valid HS clock, as verified with an oscilloscope. This
results in the downstream panel or bridge receiving invalid video data.
Setting clkdiv_clk_en in such cases resolves the HS clock problem and
the downstream device can display the video correctly.

Signed-off-by: Carter McBride <carter.mcbride@nxp.com>
Acked-by: Liu Ying <victor.liu@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>