LF-18065-5: arm64: dts: imx95: imx519: Add imx95-19x19-evk-imx519-ak7375.dtso
Add device tree for the Sony IMX519 image sensor together with
the Asahi Kasei Microdevices AK7375 voice coil lens actuator.

Use a fixed regulator reg_cam2p8v, which includes the GPIO from
the pcal6408 expander controlling MIPI_CSI_RST (pin A9) on the
mini-SAS connector. From the UC-873 module point of view, this
same signal is connected to POWER_EN (pin 17), which gates the
power supplies to both the sensor and the VCM.

The same regulator is used as the 2.8 V supply for both IMX519
and AK7375. When both devices are suspended, power is removed
from both.

Link: https://blog.arducam.com/downloads/schematics/UC-873_Rev.D_SCH.pdf
Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Tested-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>