MA-24245: mxc: vpu: hantro_v4l2: Handle min_queued_buffers
The daemon requires the buffer to be queued before the output on.
So hantro_v4l2 can not drop min_queued_buffers assignment,
Then it may change the seek flow, Especially the following process
in seek + source change:

1. output streamon (seek end)
2. capture streamon (source change end)
3. output buffer queue (actual seek end, due to min_queued_buffers)

Then the source change flow is not finished, need to execute an extra
capture on to finish source change flow.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>